Extract compare correlations views
