+++ service-worker.js
@@ -65,6 +65,7 @@ const APP_SHELL = [
   '/js/dashboard-widget-renderers.js',
   '/js/marker-detail-modal.js',
   '/js/light-conditions-now.js',
+  '/js/compare-correlations.js',
   '/js/views.js',
   '/js/recommendations.js',
   '/js/crypto.js',
