+++ tests/verify-modules.js
@@ -565,6 +565,9 @@
     const hasLightConditionsNowJs = sw.includes('/js/light-conditions-now.js');
     assert('Service worker caches js/light-conditions-now.js', hasLightConditionsNowJs);
 
+    const hasCompareCorrelationsJs = sw.includes('/js/compare-correlations.js');
+    assert('Service worker caches js/compare-correlations.js', hasCompareCorrelationsJs);
+
     const hasSchemaJs = sw.includes('/js/schema.js');
     assert('Service worker caches js/schema.js', hasSchemaJs);
